[AOT Refactor] jit compile runtime wrappers
[ghstack-poisoned]
