+++ torch/_functorch/aot_autograd.py
@@ -12,80 +12,42 @@ import torch.utils.dlpack
 from torch import Tensor
 from torch._dispatch.python import enable_python_dispatcher
 from torch._dynamo import compiled_autograd
-from torch._dynamo.utils import dynamo_timed, lazy_format_graph_code, preserve_rng_state
-from torch._guards import detect_fake_mode, tracing
-from torch._prims_common import CUDARngStateHelper
+from torch._dynamo.utils import dynamo_timed, preserve_rng_state
+from torch._guards import detect_fake_mode
 from torch._logging import getArtifactLogger
 from torch._subclasses import FakeTensor, FakeTensorMode
-from torch.fx.experimental.proxy_tensor import is_sym_node, make_fx
+from torch.fx.experimental.proxy_tensor import make_fx
 from torch.fx.experimental.symbolic_shapes import (
-    ShapeEnv, fx_placeholder_vals
+    ShapeEnv
 )
 from torch.utils._python_dispatch import is_traceable_wrapper_subclass
 from torch._decomp.decompositions_for_rng import PhiloxStateTracker, rng_decompositions
 from . import config
 from .partitioners import default_partition
-from torch._guards import TracingContext
 
 from ._aot_autograd.utils import (  # noqa: F401
-    strict_zip, _get_symint_hints, create_tree_flattened_fn,
-    KNOWN_TYPES, partial_flatten_asdict, normalize_as_list,
-    _get_autocast_states, make_boxed_func, call_func_at_runtime_with_args,
-    make_boxed_compiler, maybe_to_fresh_input,
+    strict_zip,
+    create_tree_flattened_fn,
+    make_boxed_func,
+    make_boxed_compiler,
 )
-from ._aot_autograd.logging_utils import (  # noqa: F401
-    setup_stacktrace_preservation_hooks,
-    get_aot_graph_name,
-    get_graph_being_compiled,
-    get_aot_compilation_context,
-    track_graph_compiling,
-    describe_input,
-    format_guard_bug_msg,
-    set_model_name,
-)
-from ._aot_autograd.functional_utils import (  # noqa: F401
-    gen_alias_from_base,
-    assert_functional_graph,
-    from_functional,
-)
-from ._aot_autograd.schemas import (
-    SubclassMeta,
-    ViewAndMutationMeta,
-    TensorAlias,
-    MutationType,
-    OutputType,
-    GraphSignature,
-    AOTConfig,
-)
-from ._aot_autograd.subclass_utils import (
-    requires_subclass_dispatch,
-    unwrap_tensor_subclasses,
-    wrap_tensor_subclasses,
-)
-from ._aot_autograd.collect_metadata_analysis import (
-    run_functionalized_fw_and_collect_metadata,
-)
-from ._aot_autograd.input_output_analysis import (
-    create_graph_signature,
-)
-
+from ._aot_autograd.schemas import ViewAndMutationMeta, OutputType, GraphSignature, AOTConfig
+from ._aot_autograd.functional_utils import from_functional  # noqa: F401
+from ._aot_autograd.subclass_utils import requires_subclass_dispatch
+from ._aot_autograd.collect_metadata_analysis import run_functionalized_fw_and_collect_metadata
+from ._aot_autograd.input_output_analysis import create_graph_signature
 from ._aot_autograd.traced_function_transforms import (  # noqa: F401
     create_functional_call,
     create_joint_function,
 )
-from ._aot_autograd.runtime_wrappers import (
-    create_runtime_wrapper,
-    functionalized_rng_runtime_epilogue,
-    aot_dispatch_subclass_wrapper,
+from ._aot_autograd.logging_utils import (  # noqa: F401
+    get_aot_compilation_context,
+    get_aot_graph_name,
+    get_graph_being_compiled,
 )
-from ._aot_autograd.alias_runtime_wrappers import (
-    aot_wrapper_dedupe,
-    aot_wrapper_synthetic_base,
-)
-from ._aot_autograd.dispatch_and_compile_graph import (
-    aot_dispatch_base_graph,
-    aot_dispatch_autograd_graph,
-)
+from ._aot_autograd.alias_runtime_wrappers import aot_wrapper_dedupe, aot_wrapper_synthetic_base
+from ._aot_autograd.jit_compile_runtime_wrappers import aot_dispatch_base, aot_dispatch_autograd
+from ._aot_autograd.dispatch_and_compile_graph import aot_dispatch_base_graph, aot_dispatch_autograd_graph
 
 zip = strict_zip
 
@@ -351,634 +313,6 @@ AOT_COUNTER = itertools.count()
 
 
 aot_autograd_decompositions = {}
-
-def aot_dispatch_base(flat_fn, flat_args: List[Tensor], aot_config: AOTConfig, *, fw_metadata: ViewAndMutationMeta):
-    fw_module, updated_flat_args, maybe_subclass_meta = aot_dispatch_base_graph(
-        flat_fn, flat_args, aot_config, fw_metadata=fw_metadata)
-
-    disable_amp = torch._C._is_any_autocast_enabled()
-    context = torch._C._DisableAutocast if disable_amp else nullcontext
-
-    with context(), track_graph_compiling(aot_config, "inference"):
-        compiler = aot_config.inference_compiler if aot_config.inference_compiler is not None else aot_config.fw_compiler
-        if config.functionalize_rng_ops:
-            # Add the seed and offset as example inputs to pass to the compiler
-            fake_mode = detect_fake_mode()
-            seed, offset = CUDARngStateHelper.get_torch_state_as_tuple(fake_mode)
-            updated_flat_args.extend([seed, offset])
-
-        if tracing_context := torch._guards.TracingContext.try_get():
-            tracing_context.fw_metadata = fw_metadata \
-                if maybe_subclass_meta is None else maybe_subclass_meta.fw_metadata
-        compiled_fw = compiler(fw_module, updated_flat_args)
-
-    # This boxed_call handling happens inside create_runtime_wrapper as well.
-    # However, create_runtime_wrapper does not expect the rng offsets in the
-    # output. So, we have to create another wrapper and take out the offset. As
-    # a result, we have to account for not boxed_call compilers as well.
-    if not hasattr(compiled_fw, "_boxed_call"):
-        compiled_fw = make_boxed_func(compiled_fw)
-
-    # Create a wrapper to set up the rng functionalize bits
-    @wraps(compiled_fw)
-    def rng_functionalization_wrapper(args):
-        # args is a list because compiled_fw is boxed_call
-        if fw_metadata.is_rng_op_functionalized:
-            # Add the seed and offset to args
-            seed, offset = CUDARngStateHelper.get_torch_state_as_tuple()
-            args.extend([seed, offset])
-            out = compiled_fw(args)
-            out = functionalized_rng_runtime_epilogue(fw_metadata, out)
-            return out
-        else:
-            return compiled_fw(args)
-
-    if maybe_subclass_meta is not None:
-        compiled_fw_func = aot_dispatch_subclass_wrapper(
-            rng_functionalization_wrapper, subclass_metas=fw_metadata.subclass_fw_graph_out_meta, num_fw_outs_saved_for_bw=None)
-    else:
-        compiled_fw_func = rng_functionalization_wrapper
-
-    if not hasattr(compiled_fw_func, "_boxed_call"):
-        compiled_fw_func = make_boxed_func(compiled_fw_func)
-
-    compiled_fn = create_runtime_wrapper(
-        compiled_fw_func,
-        runtime_metadata=fw_metadata,
-        indices_of_inps_to_detach=[],
-        trace_joint=False,
-        keep_input_mutations=aot_config.keep_inference_input_mutations,
-        disable_amp=disable_amp
-    )
-
-    return compiled_fn
-
-
-def aot_dispatch_autograd(flat_fn, flat_args: List[Any], aot_config: AOTConfig, *, fw_metadata: ViewAndMutationMeta):
-    fx_g, joint_inputs, maybe_subclass_meta = aot_dispatch_autograd_graph(flat_fn, flat_args, aot_config, fw_metadata=fw_metadata)
-
-    # Copied from aot_dispatch_autograd_graph.
-    traced_tangents = pytree.tree_map(
-        lambda x: x.detach().contiguous() if isinstance(x, Tensor) else x,
-        fw_metadata.traced_tangents,
-    )
-    disable_amp = torch._C._is_any_autocast_enabled()
-
-    if aot_config.enable_log:
-        aot_joint_log.info("%s", lazy_format_graph_code("Joint graph", fx_g, aot_config.aot_id))
-
-    with torch.no_grad():
-        inner_meta = fw_metadata if maybe_subclass_meta is None else maybe_subclass_meta.fw_metadata
-        with track_graph_compiling(aot_config, "joint"):
-            # See Note: [Partitioner handling for Subclasses, Part 1]
-            num_inner_fwd_outputs = (
-                inner_meta.num_mutated_inp_runtime_indices
-                + inner_meta.num_outputs
-                + inner_meta.num_intermediate_bases
-                + inner_meta.num_outputs_rng_offset
-            )
-            fw_module, bw_module = aot_config.partition_fn(
-                fx_g, joint_inputs, num_fwd_outputs=num_inner_fwd_outputs
-            )
-            fw_outs = next(n for n in fw_module.graph.nodes if n.op == "output").args[0]
-            # we only need to bookkeep the symints that are saved for bw, not any symints
-            # the user forward might have returned in its own output
-            fw_outs_saved_for_bw = fw_outs[num_inner_fwd_outputs:]
-            num_fw_outs_saved_for_bw = len(fw_outs_saved_for_bw)
-            symint_outs_saved_for_bw = [
-                n for n in fw_outs_saved_for_bw if is_sym_node(n)
-            ]
-            fw_metadata.num_symints_saved_for_bw = len(symint_outs_saved_for_bw)
-            inner_meta.num_symints_saved_for_bw = len(symint_outs_saved_for_bw)
-            _num_symints_saved_for_bw = len(symint_outs_saved_for_bw)
-
-        # Note [Detaching inputs that never need gradients]
-        # See https://github.com/pytorch/pytorch/issues/97745
-        # Suppose we have a function like this that we want to compile:
-        #
-        # def f(x, y):
-        #     return torch.mul(x, y.detach())
-        #
-        # What gradients should we compute for x and y?
-        # By default, AOTAutograd will compute a gradient for **every** input that requires gradients,
-        # and so we'll compute:
-        #    x_grad_input = y
-        #    y_grad_input = None
-        # Does this preserve the semantics of eager mode?
-        # Unfortunately, no.
-        # Doing the above will cause autograd to **continue** to backprop the autograd tape
-        # that was generated from constructing y.
-        #
-        # This is **different** from what would have happened in eager mode.
-        # In eager mode, if we backprop through the output of this function, autograd will only traverse
-        # the bit of the autograd tape corresponding to "x".
-        # In particular, if a user had previously backpropped through y's autograd tape,
-        # And then they try to backprop through the output of the above function,
-        # then we'll hit the dreaded "Trying to backward through the graph a second time" error.
-        #
-        # You might think: If autograd sees that a gradient is None, shouldn't it stop early,
-        # instead of continuing the backprop through the ancestors of that node in the graph?
-        #
-        # Autograd has two passes:
-        # (1) a first pass that traverses the autograd graph and figures out which nodes need to be executed
-        # (2) a second pass that actually goes ahead and executes each node when it becomes ready,
-        #     propagating gradients
-        # By the time we're executing a node and we see that it produces a None, the set of nodes to execute
-        # is already locked-in.
-        #
-        # The fix: instead, we can recognize statically that the graph we're compiling will never contribute
-        # gradients to y, and prevent autograd from trying to traverse y's autograd tape at all.
-        # We can do this by manually detach'ing y before sending it through the `CompiledFunction`.
-        #
-        # Note that this solution is not bulletproof.
-        # It's possible to construct a case where eager may or may not have have tried to autograd through y,
-        # depending on the actual grad_outputs that were passed in during the backward.
-        # There is no easy fix for this: the simplest fix would be to run with `retain_graph=True`,
-        # allowing autograd to re-use the graph.
-        #
-        # An example of this case is:
-        # def f(x):
-        #     return x.detach() * 2, x * 3
-        # If we were to only backprop through outs[0], in eager, we would stop
-        # If we backward only on the first output, we shouldn't send a grad through x.
-        # But the custom autograd function doesn't know that: it will materialize zero grads for x * 3
-        # and we will end up with a zero grad at x.
-        # If we later backprop through the second output, this will also require backprop'ing through x.
-        # Meaning we'll need to use `retain_graph=True` to be able to backprop through x the second time.
-        _indices_of_inps_to_detach = []
-        bw_outs = next(n for n in bw_module.graph.nodes if n.op == "output").args[0]
-
-        # TODO: we should apply the below "detach inputs if their gradients are statically known to be None"
-        # optimization even if we have subclass inputs/outputs (we do not handle this today).
-        # Computing which our our inputs get None gradients is a bit more complicated,
-        # if any of our inputs are subclasses. Why?
-        # (a) we need to make sure that we call .detach() on the input subclasses, since autograd sees subclasses.
-        # (b) The grad_outputs that we AOT computed in our backward graph are the desugared tensor tensors,
-        #     so we need to figure out which subclass fw inputs they map to.
-        if maybe_subclass_meta is None:
-            assert len(bw_outs) == len(fw_metadata.input_info) + inner_meta.num_outputs_rng_offset
-            for i, (bw_out) in enumerate(bw_outs):
-                if bw_out is None:
-                    _indices_of_inps_to_detach.append(i)
-
-        if aot_config.enable_log:
-            aot_graphs_log.info("%s", lazy_format_graph_code("Forward graph", fw_module, aot_config.aot_id))
-            aot_graphs_log.info("%s", lazy_format_graph_code("Backward graph", bw_module, aot_config.aot_id))
-
-        with track_graph_compiling(aot_config, "forward"):
-            # flat_args at this point might still be subclasses-
-            # make sure to pass the unwrapped fake tensors into the compiler!
-            adjusted_flat_args = joint_inputs[0]
-            if config.functionalize_rng_ops:
-                # Update example inputs for the fw_compiler
-                fake_mode = detect_fake_mode()
-                seed, offset = CUDARngStateHelper.get_torch_state_as_tuple(fake_mode)
-                adjusted_flat_args.extend([seed, offset])
-                # We are not clearing flat_args here because
-                # 1) There is a check in the debug compiler at the end
-                # 2) It does not matter as these are fake tensors
-
-            if tracing_context := torch._guards.TracingContext.try_get():
-                tracing_context.fw_metadata = inner_meta
-
-            with TracingContext.report_output_strides() as fwd_output_strides:
-                compiled_fw_func = aot_config.fw_compiler(
-                    fw_module, adjusted_flat_args
-                )
-            if not hasattr(compiled_fw_func, "_boxed_call"):
-                compiled_fw_func = make_boxed_func(compiled_fw_func)
-
-            if maybe_subclass_meta is not None:
-                # Why do we need to pass in num_fw_outs_saved_for_bw?
-                # See Note: [Partitioner handling for Subclasses, Part 2]
-                compiled_fw_func = aot_dispatch_subclass_wrapper(
-                    compiled_fw_func,
-                    subclass_metas=fw_metadata.subclass_fw_graph_out_meta,
-                    num_fw_outs_saved_for_bw=num_fw_outs_saved_for_bw
-                )
-                if not hasattr(compiled_fw_func, "_boxed_call"):
-                    compiled_fw_func = make_boxed_func(compiled_fw_func)
-
-        # NB: It's important to compile backwards ahead of time, as this may
-        # add extra guards which we need to apply to the Dynamo cache at
-        # forwards
-        with track_graph_compiling(aot_config, "backward"):
-            placeholder_list = fx_placeholder_vals(bw_module)
-
-            forward_saved_for_backwards_strides = None
-            if fwd_output_strides is not None:
-                forward_saved_for_backwards_strides = fwd_output_strides[inner_meta.tensors_saved_for_backwards_slice]
-
-            # saved activations can have different stride to eager if
-            # the compiler does layout optimization. We should restride the
-            # tensor passed in for compiling the backward graph using the
-            # saved tensor's stride.
-            for i in range(len(placeholder_list)):
-                ph_arg = placeholder_list[i]
-                if not isinstance(ph_arg, torch.Tensor):
-                    continue
-
-                if forward_saved_for_backwards_strides is None:
-                    continue
-
-                real_stride = None
-                # Per all_args calling convention
-                j = i - len(symint_outs_saved_for_bw)
-                if 0 <= j < len(forward_saved_for_backwards_strides):
-                    real_stride = forward_saved_for_backwards_strides[j]
-                if real_stride is None:
-                    continue
-
-                # Comparing ph_arg.stride() with real_stride directly may
-                # cause dynamic dimensions in ph_arg being specialized to static
-                # value. Using the hints to avoid that.
-                if _get_symint_hints(ph_arg.stride()) != real_stride:
-                    # Note that here we use the stride of the real tensor to
-                    # restride a FakeTensor. This does not cause trouble
-                    # for dynamic shape since this code path only get
-                    # executed if layout optimization is enabled. And we
-                    # disable layout optimization for dynamic shape right
-                    # now.
-                    #
-                    # A solution that decide stride order based on real
-                    # tensor's stride and then apply that stride order to
-                    # the FakeTensor does not work smoothly since some
-                    # tensor's layout is not 'dense'. E.g. mixnet_l has a
-                    # tensor with size [8, 64, 112, 112] and strides
-                    # (2408448, 1, 21504, 192). The solution mentioned will
-                    # decide a stride of (802816, 1, 7168, 64) for this
-                    # tensor which is wrong.
-                    placeholder_list[i] = ph_arg.as_strided(ph_arg.size(), real_stride)
-
-            compiled_bw_func = None
-            if len(symint_outs_saved_for_bw):
-                context = torch._C._DisableAutocast if disable_amp else nullcontext
-                with context():
-                    try:
-                        compiled_bw_func = aot_config.bw_compiler(
-                            bw_module, placeholder_list
-                        )
-                    except Exception:
-                        log.warning(
-                            "failed to eagerly compile backwards for dynamic, suppressing in case backwards not needed",
-                            exc_info=True
-                        )
-
-    saved_context = TracingContext.try_get()
-
-    class CompiledFunction(torch.autograd.Function):
-        compiled_fw = compiled_fw_func
-        compiled_bw = compiled_bw_func
-        metadata = fw_metadata
-        maybe_subclass_metadata: Optional[SubclassMeta] = maybe_subclass_meta
-        num_symints_saved_for_bw = _num_symints_saved_for_bw
-
-        @staticmethod
-        def _compiled_autograd_key(ctx):
-            return (aot_config.aot_id, *ctx.symints)
-
-        @staticmethod
-        def forward(ctx, *deduped_flat_tensor_args):
-            args = deduped_flat_tensor_args
-
-            marked_dirty_inps = []
-            for i in fw_metadata.mutated_graph_handled_indices:
-                ctx.mark_dirty(deduped_flat_tensor_args[i])
-                marked_dirty_inps.append(deduped_flat_tensor_args[i])
-
-            if CompiledFunction.metadata.is_rng_op_functionalized:
-                # Add the seed and offset to args
-                seed, offset = CUDARngStateHelper.get_torch_state_as_tuple()
-                args = (*args, seed, offset)
-            # There is a pretty complicated calling convention around what the compiled fw returns.
-            # The full list of outputs and their relative order is:
-            # (*mutated_inputs, *fw_outs, *fw_intermediate_bases, *saved_tensors, *saved_symints)
-            # - Note that in the synthetic bases case, mutated_inputs will correspond to an updated version
-            #   of the original view, and not the synthetic base
-            fw_outs = call_func_at_runtime_with_args(
-                CompiledFunction.compiled_fw,
-                args,
-                disable_amp=disable_amp,
-            )
-
-            num_outputs = CompiledFunction.metadata.num_outputs
-            num_outputs_aliased = CompiledFunction.metadata.num_outputs_aliased
-            num_intermediate_bases = CompiledFunction.metadata.num_intermediate_bases
-            num_symints_saved_for_bw = CompiledFunction.num_symints_saved_for_bw
-            num_mutated_runtime_inps = CompiledFunction.metadata.num_mutated_inp_runtime_indices
-            num_forward_returns = CompiledFunction.metadata.num_forward_returns
-            num_forward = CompiledFunction.metadata.num_forward
-
-            # Partitioners must put symint arguments at the end separate from tensor arguments
-            tensors_saved_for_backwards = fw_outs[
-                CompiledFunction.metadata.tensors_saved_for_backwards_slice
-            ]
-            assert all(
-                isinstance(x, torch.Tensor) for x in tensors_saved_for_backwards
-            )
-            # See Note [Detaching saved tensors in AOTAutograd]
-            ctx.save_for_backward(*(x.detach() if x._is_view() else x for x in tensors_saved_for_backwards))
-            symint_outs = fw_outs[CompiledFunction.metadata.symints_saved_for_backwards_slice]
-            assert all(
-                isinstance(x, (int, float, torch.SymInt, torch.SymFloat))
-                for x in symint_outs
-            ), str([type(x) for x in symint_outs])
-            ctx.symints = symint_outs
-
-            raw_returns = fw_outs[0:num_forward_returns]
-
-            # Wrap all autograd.Function.forward() outputs that are aliases
-            # so that autograd.Function doesn't treat them as tensors
-            if num_mutated_runtime_inps > 0:
-                for i, idx in enumerate(
-                    CompiledFunction.metadata.mutated_inp_runtime_indices
-                ):
-                    # We could make this faster by only looping over inputs with metadata-only mutations
-                    # (instead of looping over inputs with either data or metadata mutations), but there shouldn't be many.
-                    info = CompiledFunction.metadata.input_info[idx]
-                    if info.mutates_metadata and not info.mutates_data:
-                        raw_returns[i] = TensorAlias(raw_returns[i])
-
-                if config.debug_assert:
-                    user_mutated_inputs_raw = raw_returns[0:num_mutated_runtime_inps]
-                    mut_inp_infos = [
-                        x for x in CompiledFunction.metadata.input_info if x.mutates_data or x.mutates_metadata
-                    ]
-                    assert len(user_mutated_inputs_raw) == len(mut_inp_infos)
-
-            if CompiledFunction.metadata.num_unsafe_view_outputs > 0:
-                for idx in CompiledFunction.metadata.unsafe_view_out_indices:
-                    raw_return_idx = num_mutated_runtime_inps + idx
-                    o = raw_returns[raw_return_idx]
-                    raw_returns[raw_return_idx] = torch.ops.aten._unsafe_view(o, o.shape)
-
-            if num_outputs_aliased > 0:
-                for idx in CompiledFunction.metadata.aliased_out_indices:
-                    raw_return_idx = num_mutated_runtime_inps + idx
-                    raw_returns[raw_return_idx] = TensorAlias(raw_returns[raw_return_idx])
-
-                if config.debug_assert:
-                    intermediates_raw = raw_returns[num_mutated_runtime_inps + num_outputs:]
-                    assert not any(isinstance(x, TensorAlias) for x in intermediates_raw)
-
-            # invariant: intermediate bases always require gradients, so we don't have to
-            # consider marking them as non-differentiable.
-            raw_returns_not_including_intermediate_bases = raw_returns[:num_mutated_runtime_inps + num_outputs]
-            raw_returns_meta = (
-                [
-                    x for x in CompiledFunction.metadata.input_info
-                    if x.mutation_type == MutationType.MUTATED_OUT_GRAPH
-                ] + CompiledFunction.metadata.output_info
-            )
-
-            fw_outs_not_requiring_grad = [
-                x
-                for (i, x) in enumerate(raw_returns_not_including_intermediate_bases)
-                if isinstance(x, torch.Tensor)
-                and not raw_returns_meta[i].requires_grad
-            ]
-            ctx.mark_non_differentiable(*fw_outs_not_requiring_grad)
-            ctx._materialize_non_diff_grads = False
-
-            functionalized_rng_runtime_epilogue(
-                CompiledFunction.metadata,
-                fw_outs[num_forward_returns:num_forward],
-                return_new_outs=False
-            )
-            return tuple(raw_returns) + tuple(marked_dirty_inps)
-
-        @staticmethod
-        def backward(ctx, *flat_args):
-            # Calling convention: we expect a grad_out passed to the backward:
-            # - for every output of the fw that does *not* alias an input or graph intermediate
-            # - for every updated_input generated by the fw that does *not* alias an input (aka only data-mutations)
-            # - for every graph intermediate that we need to use to generate an output later.
-            # The other outputs in the autograd.Function.forward that do *not* show up in the backward include:
-            # - outputs that alias inputs or graph intermediates
-            # - updated inputs due to metadata-only mutations.
-            # We need to return them in the forward, but ensure that they all do not get gradients in the backward,
-            # and we filter them out here before passing the remaining grad_outputs into the compiled backward.
-            num_intermediate_bases = CompiledFunction.metadata.num_intermediate_bases
-            num_graph_handled_inputs = CompiledFunction.metadata.num_mutated_graph_handled_indices
-            num_mutated_runtime_inps = CompiledFunction.metadata.num_mutated_inp_runtime_indices
-            expected_grad_outs = (
-                CompiledFunction.metadata.num_outputs + num_mutated_runtime_inps + num_intermediate_bases
-            )
-
-            if num_graph_handled_inputs > 0:
-                flat_args = flat_args[:-num_graph_handled_inputs]
-            assert len(flat_args) == expected_grad_outs
-            out_info = CompiledFunction.metadata.output_info
-
-            inp_tangents, out_tangents, intermediate_base_tangents = (
-                flat_args[0:num_mutated_runtime_inps],
-                flat_args[num_mutated_runtime_inps:num_mutated_runtime_inps + CompiledFunction.metadata.num_outputs],
-                flat_args[num_mutated_runtime_inps + CompiledFunction.metadata.num_outputs:],
-            )
-            # input_info contains info on *every* input,
-            # But in the backward(), we are only given grad outputs for every mutated input
-            # We then need to filter out the grad outputs that correspond to metadata-only mutations or don't require grad
-            input_info = CompiledFunction.metadata.input_info
-            inp_tangents_filtered = [
-                x
-                for x, info_idx in zip(inp_tangents, CompiledFunction.metadata.mutated_inp_runtime_indices)
-                if input_info[info_idx].mutates_data and input_info[info_idx].requires_grad
-            ]
-            # We also need to filter out grad outputs that correspond to outputs aliasing inputs/intermediates
-            out_tangents_filtered = [
-                x
-                for x, info in zip(out_tangents, out_info)
-                if info.output_type in [OutputType.non_alias, OutputType.unsafe_view_alias, OutputType.custom_function_view]
-                and issubclass(info.raw_type, torch.Tensor)
-                and info.requires_grad
-            ]
-            # intermediate bases always require gradients, and always participate in the backward graph.
-            flat_bw_args_with_grads = [*inp_tangents_filtered, *out_tangents_filtered, *intermediate_base_tangents]
-            num_flat_bw_args_with_grads = len(flat_bw_args_with_grads)
-
-            # sanity asserts
-            # metadata_only_inps = [
-            #     x for x, info_idx in zip(inp_tangents, mutated_inp_indices)
-            #     if not input_info[info_idx].mutates_data
-            # ]
-            # aliased_outputs = [
-            #     x for x, info in zip(out_tangents, out_info) if info.output_type != OutputType.non_alias]
-            # assert all(x is None for x in metadata_only_inps)
-            # assert all(x is None for x in aliased_outputs)
-
-            rng_args = []
-            if CompiledFunction.metadata.is_rng_op_functionalized:
-                # Add the seed and offset to args
-                rng_args = CUDARngStateHelper.get_torch_state_as_tuple()
-
-            all_args = [
-                *ctx.symints,
-                *ctx.saved_tensors,
-                *flat_bw_args_with_grads,
-                *rng_args
-            ]
-            del flat_bw_args_with_grads
-
-            tangents_start_idx = len(all_args) - num_flat_bw_args_with_grads - len(rng_args)
-            tangents_end_idx = len(all_args) - len(rng_args)
-
-            # Note: [AOTAutograd Backward Guards]
-            # During AOTDispatch, we eagerly create and trace out a joint fw-bw graph.
-            # Doing so requires us to "guess" about some of the metadata of our grad_outputs.
-            #
-            # In particular: if an output to the forward is a plain tensor or a subclass,
-            # its corresponding grad_output in the backward **may or may not** be
-            # a plain tensor or a subclass. The main cases are:
-            # (1) If an output is a plain tensor, its grad_out will also be a plain tensor,
-            #     *unless* the output is used in some subclass compute later in the forward graph,
-            #     which will cause its grad_output to become a subclass
-            # (2) If an output is a subclass, its grad_out will also be a subclass,
-            #     *unless* the output of the forward did not actually participate in the gradient computation,
-            #     in which case autograd will insert a plain tensor of zeros for the grad_output.
-            #     We could avoid this case with `torch.autograd.Function.set_materialize_grads`,
-            #     although this is not turned on today in AOTAutgrad and would require more work.
-            #
-            # Today, we make a guess on subclass-ness based on the above examples,
-            # and hard-error in the backward if we guessed wrong.
-            #
-            # In the future, we should add backward guards that would allow us to
-            # properly handle this case instead of erroring: we would need to retrace the backward graph,
-            # since we might produce an entirely different trace if our grad_outputs are subclass or not.
-            assert len(CompiledFunction.metadata.output_types) == num_flat_bw_args_with_grads
-            grad_output_types = [type(x) for x in all_args[-num_flat_bw_args_with_grads:]]
-            # In general, we can add more asserts/guards here for when we partitioned
-            # with incorrect assumptions about the grad_outputs.
-            # Normalize FakeTensor -> torch.Tensor
-            # - during tracing our types are FakeTensor
-            # - at runtime in the backward our types are torch.Tensor...
-            # - unless we're running compiled backward, in which case they are also FakeTensor
-            grad_output_types_ = [torch.Tensor if x is FakeTensor else x for x in grad_output_types]
-            assert grad_output_types_ == CompiledFunction.metadata.output_types, f"""\
-We incorrectly attempted to compile the backward with incorrect subclass metadata.
-If you run into this error, please file an issue.
-Expected grad_output types: {str(CompiledFunction.metadata.output_types)}
-Got grad_output types: {str(grad_output_types)}"""
-
-            # TODO: figure out how to refactor the backward properly so I can use aot_dispatch_subclass_wrapper() here.
-            if CompiledFunction.maybe_subclass_metadata is not None:
-                # Get the number of tangents after unwrapping
-                len_tangents = len(unwrap_tensor_subclasses(
-                    all_args[tangents_start_idx: tangents_end_idx], is_joint_structure=False
-                ))
-                all_args = unwrap_tensor_subclasses(all_args, is_joint_structure=False)
-                tangents_start_idx = len(all_args) - len_tangents - len(rng_args)
-                tangents_end_idx = tangents_start_idx + len_tangents
-
-            # Make the tangents contiguous. Note that we must do this after subclass desugaring
-            # because inputs to inductor have to be contiguous
-            all_args = [
-                t.contiguous() if tangents_start_idx <= i < tangents_end_idx else t
-                for i, t in enumerate(all_args)
-            ]
-
-            def call_compiled_backward():
-                if ctx._is_compiled_autograd_tracing():
-                    # For compiled autograd, run raw FX graph so that it can be inlined into the larger graph
-                    symints = ctx._get_compiled_autograd_symints()
-                    assert len(symints) == len(ctx.symints)
-                    all_args[:len(symints)] = symints
-                    context = torch._C._DisableAutocast if disable_amp else nullcontext
-                    with context():
-                        out = normalize_as_list(bw_module(*all_args))
-                    out = functionalized_rng_runtime_epilogue(CompiledFunction.metadata, out)
-                    return tuple(out)
-                ctx.maybe_clear_saved_tensors()
-                if CompiledFunction.compiled_bw is None:
-                    context = torch._C._DisableAutocast if disable_amp else nullcontext
-                    with tracing(saved_context), context(), track_graph_compiling(aot_config, "backward"):
-                        CompiledFunction.compiled_bw = aot_config.bw_compiler(
-                            bw_module, placeholder_list
-                        )
-
-                out = call_func_at_runtime_with_args(
-                    CompiledFunction.compiled_bw,
-                    all_args,
-                    steal_args=True,
-                    disable_amp=disable_amp,
-                )
-
-                out = functionalized_rng_runtime_epilogue(CompiledFunction.metadata, out)
-                return tuple(out)
-
-            if torch.is_grad_enabled() and any(t.requires_grad for t in all_args if isinstance(t, torch.Tensor)):
-                # Ensure that the graph is connected, and error if double backward is performed.
-                # See comment for why once_differentiable is not sufficient:
-                # https://github.com/pytorch/pytorch/pull/92348/files#r1072962107
-                class CompiledFunctionBackward(torch.autograd.Function):
-                    @staticmethod
-                    def forward(ctx, *unused_args):
-                        outs = call_compiled_backward()
-                        # TODO: figure out how to refactor the backward properly so I can use aot_dispatch_subclass_wrapper() here.
-                        if CompiledFunction.maybe_subclass_metadata is not None:
-                            outs_wrapped = wrap_tensor_subclasses(
-                                outs, subclass_metas=CompiledFunction.maybe_subclass_metadata.grad_input_metas)
-                            return outs_wrapped
-                        return outs
-
-                    @staticmethod
-                    def backward(ctx, *args):
-                        raise RuntimeError("torch.compile with aot_autograd does not currently support double backward")
-
-                CompiledFunctionBackward._compiled_autograd_key = CompiledFunction._compiled_autograd_key
-
-                # Pass args even though they're unused, so that the graph is built
-                out = CompiledFunctionBackward.apply(*all_args)
-            else:
-                out = call_compiled_backward()
-
-            # TODO: figure out how to refactor the backward properly so I can use aot_dispatch_subclass_wrapper() here.
-            if CompiledFunction.maybe_subclass_metadata is not None:
-                outs_wrapped = wrap_tensor_subclasses(
-                    out, subclass_metas=CompiledFunction.maybe_subclass_metadata.grad_input_metas)
-                return outs_wrapped
-            return out
-
-    compiled_function = create_runtime_wrapper(
-        CompiledFunction.apply,
-        runtime_metadata=fw_metadata,
-        indices_of_inps_to_detach=_indices_of_inps_to_detach,
-        trace_joint=True,
-        keep_input_mutations=aot_config.keep_inference_input_mutations,
-        disable_amp=disable_amp
-    )
-
-    if not config.debug_assert:
-        return compiled_function
-
-    flat_requires_grad = [
-        a.requires_grad if isinstance(a, Tensor) else None for a in flat_args
-    ]
-
-    @wraps(compiled_function)
-    def debug_compiled_function(*args):
-        # TODO: Check aliasing relationships
-        # TODO: Check strides for metadata mutation
-        # (NB: ideally, this logic is factored out of this function and
-        # you move these debug checks there)
-
-        # Check requires grad.  Bad case is when we compiled with
-        # requires_grad = False, but input requires_grad = True
-        # (vice versa is OK; we compute a gradient and then throw
-        # it away when it hits the input.)
-        for i, a in enumerate(args):
-            can_require_grad = flat_requires_grad[i]
-            if can_require_grad is None:
-                assert not isinstance(a, Tensor)
-            elif not can_require_grad:
-                assert not a.requires_grad, format_guard_bug_msg(
-                    aot_config,
-                    f"{describe_input(i, aot_config)} would not require grad",
-                )
-
-        return compiled_function(*args)
-
-    return debug_compiled_function
-
 
 @dynamo_timed
 def create_aot_dispatcher_function(
